Add signer API code docs introductory text
The idea was to provide just enough detail to "get" the signer api at a
glance.  Finding the right words and the right level of detail was quite
hard. But it's a start.

My feeling is that the intro could use a little bit more detail, and the
docstrings a little less.

Plus there should be examples, and public methods in specific Signer
and Key implementations need to be included

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>