Mark securesystemslib.gpg subpackage as internal
The `gpg` subpackage provides a vaguely defined API (`gpg.functions`) to
create signatures, export public keys, and verify signatures. This API
and the used formats are incompatible with the securesystemslib signer
API.

For the sake of a consistent API, the `gpg` subpackage is marked as
internal (renamed to `_gpg`) and the above mentioned functionality is
exposed via the new signer API. Replacement methods are:

- `GPGSigner.import_` (replaces `export_pubkey`)
- `GPGSigner.sign`
- `GPGKey.verify_signature`

Note that public key and signature formats also change, in order to match
`Key` and `Signature` interfaces. This means:

- signature field `signature` is renamed to `sig`
- public key fields `type`, `method` and `hashes` are replaced by
  `keytype` and `scheme` fields, and
- public keys no longer include `subkeys` or key expiration infos. This
  means that the signature verification function no longer needs to
  decide, if a key is authorized or valid to verify a given signature.

See discussion for context:
https://github.com/secure-systems-lab/securesystemslib/pull/488#issuecomment-1398169012
https://github.com/secure-systems-lab/securesystemslib/pull/488#issuecomment-1400190844

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>