Refactor scheme handling in Azure and GCP signers
Refactor constructors and helpers in AzureSigner and GCPSigner to
explicitly validate supported keytypes and schemes early on, to not rely
on implicit keytype/scheme validation helpers, that parse the scheme.

The advantage is that validation is more explicit, and that we can use a
more general parsing helper, which supports more keytype/scheme pairs,
than the individual signer does.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>